chore: Move the security schema for the controller to its own file
