+++ app/Http/Controllers/Apis/Protected/Summit/OAuth2SpeakerOrganizationalRoleApiController.php
@@ -22,25 +22,6 @@ use utils\PagingResponse;
 use Illuminate\Support\Facades\Request;
 use OpenApi\Attributes as OA;
 
-
-#[OA\SecurityScheme(
-        type: 'oauth2',
-        securityScheme: 'speaker_organizational_role_oauth2',
-        flows: [
-            new OA\Flow(
-                authorizationUrl: L5_SWAGGER_CONST_AUTH_URL,
-                tokenUrl: L5_SWAGGER_CONST_TOKEN_URL,
-                flow: 'authorizationCode',
-                scopes: [
-                    SummitScopes::ReadAllSummitData => 'Read All Summit Data',
-                    SummitScopes::ReadSummitData => 'Read Summit Data',
-                ],
-            ),
-        ],
-    )
-]
-class SpeakerOrganizationalRoleAuthSchema{}
-
 /**
  * Class OAuth2SpeakerOrganizationalRoleApiController
  * @package App\Http\Controllers
