+++ app/Swagger/Security/SpeakerOrganizationalRoleAuthSchema.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Swagger\schemas;
+
+use App\Security\SummitScopes;
+use OpenApi\Attributes as OA;
+
+
+#[OA\SecurityScheme(
+        type: 'oauth2',
+        securityScheme: 'speaker_organizational_role_oauth2',
+        flows: [
+            new OA\Flow(
+                authorizationUrl: L5_SWAGGER_CONST_AUTH_URL,
+                tokenUrl: L5_SWAGGER_CONST_TOKEN_URL,
+                flow: 'authorizationCode',
+                scopes: [
+                    SummitScopes::ReadAllSummitData => 'Read All Summit Data',
+                    SummitScopes::ReadSummitData => 'Read Summit Data',
+                ],
+            ),
+        ],
+    )
+]
+class SpeakerOrganizationalRoleAuthSchema{}
